+++ src/main/java/com/skl/community/community/service/QuestionService.java
@@ -76,7 +76,7 @@ public class QuestionService {
 
   }
 
-  public PaginationDTO list(Integer userId, Integer page, Integer size) {
+  public PaginationDTO list(Long userId, Integer page, Integer size) {
     PaginationDTO paginationDTO = new PaginationDTO();
     Integer totalPage;
     QuestionExample questionExample = new QuestionExample();
@@ -123,7 +123,7 @@ public class QuestionService {
     return paginationDTO;
   }
 
-  public QuestionDTO getById(Integer id) {
+  public QuestionDTO getById(Long id) {
     Question question = questionMapper.selectByPrimaryKey(id);
     if (question==null){
       throw new CommunityException(CommunityErrorCode.QUESTION_NOT_FOUND);
@@ -164,7 +164,7 @@ public class QuestionService {
     }
   }
 
-  public void incView(Integer id) {
+  public void incView(Long id) {
     Question question = new Question();
     question.setId(id);
     question.setViewCount(1);
